fix(collision): FUN_0014df70 fog-zone pg_surf read as signed 16-bit word
c10 (Silent Cartographer) scenario_load halted at tag_groups.c:3089
"#65536 is not a valid <unknown> index in [#0,#1)". Root-caused to the
fog/atmosphere-zone test in FUN_0014df70: pg_surf was read as a 32-bit
dword (`*(int*)(pg+0x24)`) where the original reads a signed 16-bit word
(`movsx eax, word ptr [eax+0x24]` @0x14e23b). The pg element's +0x24 is a
16-bit surface index and +0x26 is an adjacent field; the dword read
combined them, yielding 0x00010000 (65536) when +0x26=1, overflowing the
count-1 fog block at scen+0x190 and firing the assert. The original's
16-bit read maxes at 0x7fff and never overflows, which is why retail c10
fog never crashes.

Fix: `short pg_surf` + `*(short*)(pg+0x24)`. Rebuilt code now emits
`movswl 0x24(%eax),%eax`, byte-identical to the original. Distinct from
the earlier main-menu fix (tag_groups:3086, render path +0x1bd) — same
function, different site (+0x2b0) and path.

Verified: instruction match vs cachebeta.xbe; VC71 85.6% (unchanged
structural ceiling); box-confirmed — c10 loads without the assert.

+++ src/halo/physics/collision_usage.c
@@ -851,7 +851,7 @@ bool FUN_0014df70(uint32_t collision_flags, float *origin, float *direction,
   int object_handle;
   char bsp_hit;
   short pg_idx;
-  int pg_surf;
+  short pg_surf;
 
   result = 0;
 
@@ -1019,7 +1019,12 @@ bool FUN_0014df70(uint32_t collision_flags, float *origin, float *direction,
       if (*(short *)((char *)pg_list + 2) != -1) {
         pg_idx = *(short *)pg_list;
         pg = tag_block_get_element((char *)scen + 0x184, (int)pg_idx, 0x28);
-        pg_surf = *(int *)((char *)pg + 0x24);
+        /* Original reads pg_surf as a SIGNED 16-bit word (movsx eax,word ptr
+         * [eax+0x24] @0x14e23b), not a dword. Reading *(int*) here combines the
+         * +0x24 surface index with the +0x26 field: when +0x24=0 and +0x26=1 the
+         * dword is 0x00010000 (65536), which overflows the count-1 fog block at
+         * scen+0x190 and halts at tag_groups.c:3089. See feedback_check_disasm. */
+        pg_surf = *(short *)((char *)pg + 0x24);
         fog_tag =
           tag_block_get_element((char *)scen + 0x190, (int)pg_surf, 0x88);
         fog = tag_get(0x666f6720, *(int *)((char *)fog_tag + 0x2c));
